requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
aiosqlite
python-dotenv
img2pdf
Pillow
torch
opencv-python-headless
einops
numpy
sqlalchemy
alembic
